Respect hidden file convention when loading arenas
[Apache Subversion](http://en.wikipedia.org/wiki/Apache_Subversion) is a revision control system.
Its normal operating procedure is to add a hidden .svn file to every folder to be versioned.

Before:
```
plugins/
  CaptureThePoints.jar
  CaptureThePoints/
    Arenas/
```

After:
```
.svn/
plugins/
  .svn/
  CaptureThePoints.jar
  CaptureThePoints/
    .svn/
    Arenas/
      .svn/
```

The structure of the `.svn` folder is as follows:
```
.svn/
  dir-prop-base
  entries
  prop-base/
  props/
  text-base/
  tmp/
```

CaptureThePoints does not respect the linux/unix convention that all files that start with a `.` are hidden files.
Hidden files should not be accessed unless your application is looking for them specifically.

Without this patch, CaptureThePoints takes the files in `.svn` folders and attempts to make arenas from them.

This is a compatability patch. While Subversion is the example, there may be hidden files on other systems that CaptureThePoints should not be accessing.